Move RecursivelyRemoveDirectory out of utils:files
This chips away at making utils:files smaller.

Bug: b/536109448
