+++ base/cvd/cuttlefish/common/libs/utils/files.h
@@ -65,7 +65,6 @@ Result<void> ChangeGroup(const std::string& path,
                          const std::string& group_name);
 bool CanAccess(const std::string& path, int mode);
 Result<bool> IsDirectoryEmpty(const std::string& path);
-Result<void> RecursivelyRemoveDirectory(const std::string& path);
 bool Copy(const std::string& from, const std::string& to);
 off_t FileSize(const std::string& path);
 Result<void> RemoveFile(const std::string& file);
