+++ base/cvd/cuttlefish/host/commands/cvd/cache/BUILD.bazel
@@ -14,6 +14,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/utils:disk_usage",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/result",
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/log:check",
