+++ base/cvd/cuttlefish/host/commands/cvd/cache/cache.cpp
@@ -33,6 +33,7 @@
 
 #include "cuttlefish/common/libs/utils/disk_usage.h"
 #include "cuttlefish/common/libs/utils/files.h"
+#include "cuttlefish/files/recursively_remove_directory.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
