+++ base/cvd/cuttlefish/host/commands/cvd/instances/BUILD.bazel
@@ -61,6 +61,7 @@ cf_cc_library(
     deps = [
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/host/commands/cvd/instances",
         "//cuttlefish/posix:strerror",
         "//libbase",
@@ -152,6 +153,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:json",
         "//cuttlefish/common/libs/utils:signals",
         "//cuttlefish/common/libs/utils:users",
+        "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/host/commands/cvd/cli:command_request",
         "//cuttlefish/host/commands/cvd/cli:utils",
         "//cuttlefish/host/commands/cvd/cli/commands:host_tool_target",
