+++ base/cvd/cuttlefish/host/commands/snapshot_util_cvd/BUILD.bazel
@@ -22,6 +22,7 @@ cf_cc_binary(
         "//cuttlefish/common/libs/utils:json",
         "//cuttlefish/common/libs/utils:tee_logging",
         "//cuttlefish/common/libs/utils:users",
+        "//cuttlefish/files:recursively_remove_directory",
         "//cuttlefish/flag_parser",
         "//cuttlefish/host/libs/command_util",
         "//cuttlefish/host/libs/command_util:libcuttlefish_run_cvd_proto",
